# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School API",
        "description": "API for managing school data",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/api/v1/students/{student}/skill-ratings": {
            "get": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "List skill ratings for a student",
                "description": "Returns the student's skill ratings filtered by session and term.",
                "operationId": "bf112465892bf1275c907d9e1c9c399d",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "session_id",
                        "in": "query",
                        "description": "Session ID (defaults to student's current session)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "term_id",
                        "in": "query",
                        "description": "Term ID (defaults to student's current term)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List returned successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "Create a skill rating for a student",
                "description": "Records a new skill rating for the given session and term.",
                "operationId": "988de1260d9a663f5ad04d9c42bcf3f3",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "skill_type_id",
                                    "rating_value"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "term_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "skill_type_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "rating_value": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skill rating created"
                    },
                    "409": {
                        "description": "Duplicate rating"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/students/{student}/skill-types": {
            "get": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "List available skill types for a student",
                "description": "Returns skill types scoped to the student's school.",
                "operationId": "097d55a2c97d99a3bc7dd4b46fe71e08",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill types returned"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/students/{student}/skill-ratings/{skillRating}": {
            "put": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "Update a student's skill rating",
                "description": "Updates session, term, skill type, or rating value for an existing record.",
                "operationId": "221d6dfba0ad1de187bb017bdd7c2aec",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "skillRating",
                        "in": "path",
                        "description": "Skill rating ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "session_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "term_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "skill_type_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "rating_value": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill rating updated"
                    },
                    "404": {
                        "description": "Skill rating not found for student"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "school-v1.4"
                ],
                "summary": "Delete a student's skill rating",
                "operationId": "5c759591a331b40b33b1935f8a9207fd",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "skillRating",
                        "in": "path",
                        "description": "Skill rating ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Skill rating deleted"
                    },
                    "404": {
                        "description": "Skill rating not found for student"
                    },
                    "422": {
                        "description": "Term closed for updates"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}